Use Google account ID in OAuth adapter
As recommended by Google, the account ID is preferred to the email
address as the identifier since email addresses can change. This allows
maintaining the account linking after we change our email domain to
endlessos.org. Unfortunately, it will make all the existing links
invalid and phabricator only allows a single Google auth token to be
linked to an account. The existing links will need to be changed in the
database when this lands so users aren't locked out of their phabricator
accounts.

https://phabricator.endlessm.com/T29811